+++ assets/eip-7619/pqclean-implementation/Makefile
@@ -3,9 +3,9 @@
 
 LIB=libfalcon-512_clean.a
 
-SOURCES = codec.c common.c fft.c fpr.c keygen.c pqclean.c rng.c sign.c verify-signature.c
-OBJECTS = codec.o common.o fft.o fpr.o keygen.o pqclean.o rng.o sign.o verify-signature.o
-HEADERS = api.h fpr.h inner.h
+SOURCES = codec.c common.c fft.c normalizer.c keygen.c pqclean.c rng.c sign.c verify-signature.c
+OBJECTS = codec.o common.o fft.o normalizer.o keygen.o pqclean.o rng.o sign.o verify-signature.o
+HEADERS = api.h normalizer.h inner.h
 
 CFLAGS=-std=c99 -O2 -W -Wall -Wconversion -Wextra -Wpedantic -Wvla -Werror -Wmissing-prototypes -Wredundant-decls -I../../../common $(EXTRAFLAGS)
